refactor(spark): consolidate the ORC reader into hudi-spark-common
The six SparkXXOrcReader classes only differed in three version hooks:
- partitionedFileToPath: replaced with the existing
  sparkAdapter.getSparkPartitionedFileUtils.getPathFromPartitionedFile
  abstraction (String path on 3.3, SparkPath since 3.4)
- structTypeToAttributes: replaced with the existing
  sparkAdapter.getSchemaUtils.toAttributes abstraction (SPARK-44353
  moved StructType#toAttributes in 3.5)
- OrcColumnarBatchReader construction: the constructor gained a
  MemoryMode parameter in Spark 3.4 (off-heap support), which cannot be
  bridged by shared source, so the per-version adapters now pass a
  one-line factory function into the shared builder

SparkOrcReaderBase becomes a concrete reader with the shared read()
and build() logic; the 4.x adapters share one createOrcFileReader
override in BaseSpark4Adapter (all 4.x use the 2-arg constructor).
